+++ CHANGELOG.md
@@ -10,8 +10,9 @@ This file is used to track changes made to the project over time.
 #### Joints
 - Python bindings for continuous joints (`JointModelRUB*`) and prismatic joints (`JointModelP*`).
 
-### Fixed
+### Misc
 - Fix docs.rs build for all crates
+- Add `python` feature to workspace and all crates for building Python bindings
 
 ## [0.0.1] - 2026-01-25
 ### Added